refactor : onboarding page 마이그레이션
