--- src/app/LogoIcon.tsx
@@ -1,54 +1,0 @@
-// TODO : Icon 컴포넌트로 변경
-const LogoIcon = () => {
-  return (
-    <svg width="201" height="76" viewBox="0 0 201 76" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <g clipPath="url(#clip0_189_2661)">
-        <g filter="url(#filter0_d_189_2661)">
-          <path d="M10.5 25.22V14.9672H28.3617V62.0397H16.8838V25.22H10.5Z" fill="white" />
-          <path
-            d="M35.9047 37.8586C35.9047 30.4646 37.3233 24.6397 40.1606 20.3838C43.0408 16.1279 47.662 14 54.0243 14C60.3867 14 64.9864 16.1279 67.8237 20.3838C70.7039 24.6397 72.144 30.4646 72.144 37.8586C72.144 45.3386 70.7039 51.2066 67.8237 55.4624C64.9864 59.7183 60.3867 61.8462 54.0243 61.8462C47.662 61.8462 43.0408 59.7183 40.1606 55.4624C37.3233 51.2066 35.9047 45.3386 35.9047 37.8586ZM61.3109 37.8586C61.3109 33.5168 60.838 30.1852 59.8923 27.8638C58.9465 25.4994 56.9906 24.3172 54.0243 24.3172C51.0581 24.3172 49.1022 25.4994 48.1564 27.8638C47.2107 30.1852 46.7378 33.5168 46.7378 37.8586C46.7378 40.7818 46.9098 43.2107 47.2537 45.1452C47.5976 47.0367 48.2854 48.5843 49.3171 49.7879C50.3918 50.9486 51.9609 51.529 54.0243 51.529C56.0878 51.529 57.6354 50.9486 58.6671 49.7879C59.7418 48.5843 60.4511 47.0367 60.795 45.1452C61.1389 43.2107 61.3109 40.7818 61.3109 37.8586Z"
-            fill="white"
-          />
-          <path
-            d="M130.33 16.7728V62.0397H119.303V34.8924L109.179 62.0397H100.281L90.0924 34.8279V62.0397H79.0658V16.7728H92.0913L104.794 48.1114L117.369 16.7728H130.33Z"
-            fill="white"
-          />
-          <path
-            d="M189.523 16.7728V62.0397H178.496V34.8924L168.372 62.0397H159.474L149.286 34.8279V62.0397H138.259V16.7728H151.285L163.988 48.1114L176.562 16.7728H189.523Z"
-            fill="white"
-          />
-        </g>
-      </g>
-      <defs>
-        <filter
-          id="filter0_d_189_2661"
-          x="0.5"
-          y="4"
-          width="199.023"
-          height="68.0397"
-          filterUnits="userSpaceOnUse"
-          colorInterpolationFilters="sRGB"
-        >
-          <feFlood floodOpacity="0" result="BackgroundImageFix" />
-          <feColorMatrix
-            in="SourceAlpha"
-            type="matrix"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-            result="hardAlpha"
-          />
-          <feOffset />
-          <feGaussianBlur stdDeviation="5" />
-          <feComposite in2="hardAlpha" operator="out" />
-          <feColorMatrix type="matrix" values="0 0 0 0 0.121492 0 0 0 0 0.133128 0 0 0 0 0.237856 0 0 0 0.08 0" />
-          <feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow_189_2661" />
-          <feBlend mode="normal" in="SourceGraphic" in2="effect1_dropShadow_189_2661" result="shape" />
-        </filter>
-        <clipPath id="clip0_189_2661">
-          <rect width="200" height="76" fill="white" transform="translate(0.5)" />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-};
-
-export default LogoIcon;
